[fix] change spot lng lng type string to double

+++ app/src/main/java/casestudyteam5/it7th/hal/ac/jp/spotin/map/MapActivity.kt
@@ -86,8 +86,8 @@ class MapActivity : AppCompatActivity(), MapContract.View, OnMapReadyCallback, G
     if (map == null) return
     val list = spots.map {
       val marker = map!!.addMarker(createSpotMarkerOption(
-        lat = it.lat.toDouble(),
-        lng = it.lng.toDouble(),
+        lat = it.lat,
+        lng = it.lng,
         name = it.name))
       MapPresenter.MarkerData(it, marker)
     }
